+++ demisto_sdk/commands/common/errors.py
@@ -22,7 +22,7 @@ ALLOWED_IGNORE_ERRORS = [
     'MP106',
     'PA113', 'PA116', 'PA124', 'PA125', 'PA127', 'PA129',
     'PB104', 'PB105', 'PB106', 'PB110', 'PB111', 'PB112', 'PB114', 'PB115', 'PB116', 'PB107',
-    'RM100', 'RM102', 'RM104', 'RM106',
+    'RM100', 'RM102', 'RM104', 'RM106', 'RM108',
     'RP102', 'RP104',
     'SC100', 'SC101', 'SC105', 'SC106',
     'IM111'
@@ -54,6 +54,7 @@ ERROR_CODE = {
     "spaces_in_the_end_of_id": {'code': "BA112", 'ui_applicable': False, 'related_field': 'id'},
     "spaces_in_the_end_of_name": {'code': "BA113", 'ui_applicable': False, 'related_field': 'name'},
     "changed_pack_name": {'code': "BA114", 'ui_applicable': False, 'related_field': 'name'},
+    "file_cannot_be_deleted": {'code': "BA115", 'ui_applicable': False, 'related_field': ''},
 
     # BC - Backward Compatible
     "breaking_backwards_subtype": {'code': "BC100", 'ui_applicable': False, 'related_field': 'subtype'},
@@ -225,6 +226,7 @@ ERROR_CODE = {
                                   'related_field': '<parameter-name>.fromlicense'},
     "changed_integration_yml_fields": {'code': "IN147", "ui_applicable": False, 'related_field': 'script'},
     "parameter_is_malformed": {'code': "IN148", 'ui_applicable': False, 'related_field': 'configuration'},
+    'empty_outputs_common_paths': {'code': 'IN149', 'ui_applicable': False, 'related_field': 'contextOutput'},
 
     # IT - Incident Types
     "incident_type_integer_field": {'code': "IT100", 'ui_applicable': True, 'related_field': ''},
@@ -434,6 +436,11 @@ def error_code_decorator(func, *args, **kwargs):
 
 class Errors:
     BACKWARDS = "Possible backwards compatibility break"
+
+    @staticmethod
+    @error_code_decorator
+    def file_cannot_be_deleted(file_path: str):
+        return f"The file {file_path} cannot be deleted. Please restore the file."
 
     @staticmethod
     def suggest_fix(file_path: str, *args: Any, cmd: str = 'format') -> str:
@@ -2161,7 +2168,8 @@ class Errors:
     def changed_pack_name(original_name):
         return f'Pack folder names cannot be changed, please rename it back to {original_name}.' \
                f' If you wish to rename the pack, you can edit the name field in pack_metadata.json,' \
-               f' and the pack will be shown in the Marketplace accordingly.'
+               f' and the pack will be shown in the Marketplace accordingly.\n' \
+               f"If the file wasn't renamed, try pulling changes from master and re-run validations"
 
     @staticmethod
     @error_code_decorator
@@ -2178,7 +2186,7 @@ class Errors:
     @staticmethod
     @error_code_decorator
     def aliases_with_inner_alias(invalid_aliases: List[str]):
-        return "The following fields exist as aliases and therefore cannot contain an 'Aliases' key."\
+        return "The following fields exist as aliases and therefore cannot contain an 'Aliases' key." \
                f"\n{invalid_aliases}\n" \
                "Please remove the key from the fields or removed the fields from the other field's Aliases list."
 
@@ -2186,3 +2194,11 @@ class Errors:
     @error_code_decorator
     def missing_readme_file(location):
         return f'{location} is missing a README file'
+
+    @staticmethod
+    @error_code_decorator
+    def empty_outputs_common_paths(paths: Dict[str, List[str]], yaml_path: str):
+        commands_str = '\n'.join(f'{command}:\t' + ", ".join(outputs) for command, outputs in paths.items())
+
+        return f"The following command outputs are missing: \n{commands_str}\n" \
+               f"please type them or run demisto-sdk format -i {yaml_path}"