+++ demisto_sdk/commands/common/tests/readme_test.py
@@ -53,15 +53,16 @@ def test_is_file_valid(mocker, current, answer):
 
 @pytest.mark.parametrize("current, answer", README_INPUTS)
 def test_is_file_valid_mdx_server(mocker, current, answer):
-    readme_validator = ReadMeValidator(current)
-    valid = readme_validator.are_modules_installed_for_verify(readme_validator.content_path)
-    if not valid:
-        pytest.skip('skipping mdx server test. ' + MDX_SKIP_NPM_MESSAGE)
-        return
-    mocker.patch.dict(os.environ, {'DEMISTO_README_VALIDATION': 'yes'})
-    assert readme_validator.is_valid_file() is answer
-    assert ReadMeValidator._MDX_SERVER_PROCESS is not None
-    ReadMeValidator.stop_mdx_server()
+    ReadMeValidator.add_node_env_vars()
+    with ReadMeValidator.start_mdx_server():
+        readme_validator = ReadMeValidator(current)
+        valid = readme_validator.are_modules_installed_for_verify(readme_validator.content_path)
+        if not valid:
+            pytest.skip('skipping mdx server test. ' + MDX_SKIP_NPM_MESSAGE)
+            return
+        mocker.patch.dict(os.environ, {'DEMISTO_README_VALIDATION': 'yes'})
+        assert readme_validator.is_valid_file() is answer
+        assert ReadMeValidator._MDX_SERVER_PROCESS is not None
 
 
 def test_are_modules_installed_for_verify_false_res(tmp_path):
@@ -321,7 +322,7 @@ def test_invalid_short_file(capsys):
     assert short_readme_error in stdout
 
 
-def test_demisto_in_readme(repo):
+def test_demisto_in_integration_readme(repo):
     """
         Given
             - An integration README contains the word 'Demisto'.
@@ -345,6 +346,36 @@ def test_demisto_in_readme(repo):
         readme_validator = ReadMeValidator(integration.readme.path)
 
         assert not readme_validator.verify_demisto_in_readme_content()
+
+
+def init_readmeValidator(readme_validator, repo, readme_path):
+    readme_validator.content_path = str(repo.path)
+    readme_validator.file_path = readme_path
+
+
+def test_demisto_in_repo_readme(mocker, repo):
+    """
+        Given
+            - A repo README contains the word 'Demisto'.
+
+        When
+            - Running verify_demisto_in_readme_content.
+
+        Then
+            - Ensure that the validation not fails.
+    """
+    from pathlib import Path
+
+    readme_path = Path(repo.path) / 'README.md'
+    mocker.patch.object(ReadMeValidator, '__init__', return_value=None)
+
+    with open(readme_path, 'w') as f:
+        f.write('This checks if we have the word Demisto in the README.')
+
+    with ChangeCWD(repo.path):
+        readme_validator = ReadMeValidator()
+        init_readmeValidator(readme_validator, repo, readme_path)
+        assert readme_validator.verify_demisto_in_readme_content()
 
 
 def test_demisto_not_in_readme(repo):